Fix type of parameter to SAFEREAD in unixfork.c

+++ src/unixfork.c
@@ -52,7 +52,7 @@ static char shcom[512]; /* Here because I'm suspicious of */
                         /* large allocations on the stack */
 
 
-static inline ssize_t SAFEREAD(int f, char *b, int c)
+static inline ssize_t SAFEREAD(int f, char *b, size_t c)
 {
   ssize_t res;
   do {
